+++ Examples/Pretty/Pods/Target Support Files/SwiftGridView/SwiftGridView-dummy.m
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+@interface PodsDummy_SwiftGridView : NSObject
+@end
+@implementation PodsDummy_SwiftGridView
+@end