+++ Examples/pod_setup.sh
@@ -6,5 +6,11 @@ pod install
 echo ''
 cd ..
 
+echo 'Setting up Pretty example...'
+cd Pretty/
+pod install
+echo ''
+cd ..
+
 echo 'Done!'
 echo ''