+++ Source/SwiftGridView.swift
@@ -335,7 +335,53 @@ open class SwiftGridView : UIView, UICollectionViewDataSource, UICollectionViewD
             self.sgCollectionView.reloadItems(at: convertedPaths)
             completion?(true) // TODO: Fix!
         }
+    }
+    
+    open func indexPathForItem(at point: CGPoint) -> IndexPath? {
+        if let cvIndexPath: IndexPath = self.sgCollectionView.indexPathForItem(at: point) {
+            let convertedPath: IndexPath = self.convertCVIndexPathToSGIndexPath(cvIndexPath)
+            
+            return convertedPath
+        }
         
+        return nil
+    }
+    
+    open func indexPath(for cell: SwiftGridCell) -> IndexPath? {
+        if let cvIndexPath: IndexPath = self.sgCollectionView.indexPath(for: cell) {
+            let convertedPath: IndexPath = self.convertCVIndexPathToSGIndexPath(cvIndexPath)
+            
+            return convertedPath
+        }
+        
+        return nil
+    }
+    
+    open func cellForItem(at indexPath: IndexPath) -> SwiftGridCell? {
+        let revertedPath: IndexPath = self.reverseIndexPathConversion(indexPath)
+        let cell = self.sgCollectionView.cellForItem(at: revertedPath) as? SwiftGridCell
+        
+        return cell
+    }
+    
+    open var visibleCells: [SwiftGridCell] {
+        get {
+            let cells = self.sgCollectionView.visibleCells as! [SwiftGridCell]
+        
+            return cells
+        }
+    }
+    
+    open var indexPathsForVisibleItems: [IndexPath] {
+        get {
+            var indexPaths = [IndexPath]()
+            for indexPath in self.sgCollectionView.indexPathsForVisibleItems {
+                let convertedPath = self.convertCVIndexPathToSGIndexPath(indexPath)
+                indexPaths.append(convertedPath)
+            }
+            
+            return indexPaths
+        }
     }
     
     // FIXME: Doesn't work as intended.